pyproject.toml
[project]
name = "norweg"
version = "0.1.0"
description = "Project that will help Poles to learn norwegian from textbooks."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyascii==0.3.3",
    "anyio==4.10.0",
    "astor==0.8.1",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "cachetools==6.1.0",
    "certifi==2025.8.3",
    "chardet==5.2.0",
    "charset-normalizer==3.4.3",
    "click==8.1.8",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "contourpy==1.3.3",
    "cssselect==1.3.0",
    "cssutils==2.11.1",
    "dataclasses-json==0.6.7",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "einops==0.8.1",
    "et-xmlfile==2.0.0",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "fonttools==4.59.2",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "ftfy==6.3.1",
    "googletrans==4.0.2",
    "gputil==1.4.0",
    "greenlet==3.2.4",
    "gtts==2.5.4",
    "h11==0.16.0",
    "h2==4.3.0",
    "h5py==3.14.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "hyperframe==6.1.0",
    "idna==3.10",
    "imagesize==1.4.1",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.74",
    "langchain-openai==0.3.29",
    "langchain-text-splitters==0.3.9",
    "langdetect==1.0.9",
    "langsmith==0.4.13",
    "lxml==6.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.6",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "msgspec==0.19.0",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "numpy==2.3.2",
    "ollama==0.5.3",
    "onnx==1.17.0", # Using your working version
    "openai==1.99.6",
    "opencv-contrib-python==4.10.0.84",
    "opencv-python==4.11.0.86",
    "openpyxl==3.1.5",
    "opt-einsum==3.3.0",
    "orjson==3.11.1",
    "packaging==25.0",
    "paddleocr==3.1.0",
    "paddlepaddle-gpu==2.6.2", # Your working version
    "paddlex==3.1.3",
    "page-dewarp==0.2.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "premailer==3.10.0",
    "prettytable==3.16.0",
    "propcache==0.3.2",
    "protobuf==3.20.2", # Your working version
    "py-cpuinfo==9.0.0",
    "pyclipper==1.3.0.post6",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pyparsing==3.2.3",
    "pypdfium2==4.30.0",
    "python-dateutil==2.9.0.post0",
    "python-doctr==1.0.0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rapidfuzz==3.13.0",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "ruamel-yaml==0.18.14",
    "ruamel-yaml-clib==0.2.12",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "setuptools==80.9.0",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "tokenizers==0.21.4",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "ujson==5.10.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "validators==0.35.0",
    "wcwidth==0.2.13",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
